# pnmc-api/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PNMC Migration API
  version: 1.0.0
  summary: Contrato API inicial para migracion incremental PNMC
  description: |
    Contrato base para desacoplar integraciones sensibles del frontend actual,
    habilitando una migracion paralela sin tocar la UI existente.
servers:
  - url: http://localhost:8080
    description: Local
  - url: https://api-dev.pnmc.local
    description: Dev (referencia)
...
paths:
...
  /api/v1/map/summary:
    get:
      tags: [mapa]
      summary: Resumen agregado del mapa por capa
      parameters:
        - in: query
          name: layer
          schema:
            $ref: '#/components/schemas/MapLayer'
          description: Capa objetivo
      responses:
        '200':
          description: Resumen por departamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MapSummaryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/map/departments/{departmentCode}/drilldown:
    get:
      tags: [mapa]
      summary: Drilldown de un departamento
      parameters:
        - in: path
          name: departmentCode
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 64
          description: Nombre normalizado de departamento
      responses:
        '200':
          description: Detalle por capa del departamento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentDrilldownResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/editorial/resources:
    get:
      tags: [editorial]
      summary: Listar recursos editoriales
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - in: query
          name: section
          schema:
            type: string
        - in: query
          name: year
          schema:
            type: string
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: Recursos retornados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedEditorialResponse'
  /api/v1/editorial/resources/{resourceId}:
    get:
      tags: [editorial]
      summary: Obtener recurso editorial por id
      parameters:
        - in: path
          name: resourceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recurso encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditorialResource'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/gallery/albums:
    get:
      tags: [galeria]
      summary: Listar albumes de galeria
      parameters:
        - in: query
          name: featuredOnly
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Albumes
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/GalleryAlbum'
  /api/v1/gallery/albums/{albumId}:
    get:
      tags: [galeria]
      summary: Obtener album por id
      parameters:
        - in: path
          name: albumId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Album encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryAlbum'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/admin/data/editorial/resources:
    post:
      tags: [admin-data]
      summary: Crear o actualizar recurso editorial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditorialResourceUpsertRequest'
      responses:
        '200':
          description: Recurso persistido
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
...
  responses:
    BadRequest:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    MapLayer:
      type: string
      enum:
        - General
        - Festivales
        - Escuelas de Música
        - Mercados Musicales
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
    Pagination:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer }
        offset: { type: integer }
        total: { type: integer }
...
    EditorialResource:
      type: object
      required: [id, title]
      properties:
        id: { type: string }
        title: { type: string }
        year: { type: string }
        section: { type: string }
        sectionPath: { type: string }
        publicationType: { type: string }
        practice: { type: string }
        category: { type: string }
        subcategory: { type: string }
        author: { type: string }
        corporateAuthor: { type: string }
        credits: { type: string }
        isbn: { type: string }
        ismn: { type: string }
        formatSize: { type: string }
        pages: { type: string }
        duration: { type: string }
        regionalScope: { type: string }
        location: { type: string }
        url: { type: string }
        keywords:
          type: array
          items: { type: string }
        summary: { type: string }
        additionalFields: { type: string }
        coverText: { type: string }
        thumbnail: { type: string }
        displayAuthor: { type: string }
    PagedEditorialResponse:
      type: object
      required: [items, pagination]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EditorialResource'
        pagination:
          $ref: '#/components/schemas/Pagination'
    GalleryPhoto:
      type: object
      properties:
        id: { type: string }
        src: { type: string }
        title: { type: string }
        alt: { type: string }
        description: { type: string }
    GallerySection:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        type: { type: string }
        photos:
          type: array
          items:
            $ref: '#/components/schemas/GalleryPhoto'
    GalleryAlbum:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        category: { type: string }
        description: { type: string }
        location: { type: string }
        dateLabel: { type: string }
        featured: { type: boolean }
        cover: { type: string }
        sections:
          type: array
          items:
            $ref: '#/components/schemas/GallerySection'
    MapDepartmentSummary:
      type: object
      required: [department, records]
      properties:
        department: { type: string }
        records: { type: integer }
        festivals: { type: integer }
        schools: { type: integer }
        markets: { type: integer }
    MapSummaryResponse:
      type: object
      required: [layer, items]
      properties:
        layer:
          $ref: '#/components/schemas/MapLayer'
        items:
          type: array
          items:
            $ref: '#/components/schemas/MapDepartmentSummary'
    DepartmentDrilldownResponse:
      type: object
      required: [department, festivals, schools, markets]
      properties:
        department: { type: string }
        festivals:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              municipality: { type: string }
        schools:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              municipality: { type: string }
              students: { type: integer }
              teachers: { type: integer }
        markets:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              name: { type: string }
              municipality: { type: string }
              averageProjects: { type: integer }
              averageBuyers: { type: integer }
...
    EditorialResourceUpsertRequest:
      type: object
      required: [title]
      properties:
        id: { type: string }
        title: { type: string }
        summary: { type: string }
        category: { type: string }
        section: { type: string }
        sectionPath: { type: string }
        publicationType: { type: string }
        author: { type: string }
        corporateAuthor: { type: string }
        year: { type: string }
        url: { type: string }
        keywords:
          type: array
          items: { type: string }
...